Show username instead of ID
